refactor: share _collect_matrix util
