+++ tools/clustering.py
@@ -36,6 +36,7 @@ import pandas as pd
 from sklearn.neighbors import LocalOutlierFactor
 from sklearn.metrics import silhouette_score, pairwise_distances
 from sklearn.mixture import GaussianMixture
+from .matrix_utils import _collect_matrix
 
 # --- HDBSCAN ---
 try:
@@ -221,22 +222,6 @@ def distance_to_medoid(Z: np.ndarray, labels: np.ndarray) -> np.ndarray:
 # ======================================================
 # --------------- ПРОТОТИПЫ КЛАСТЕРОВ ------------------
 # ======================================================
-
-def _collect_matrix(panel_long: pd.DataFrame, wells: Sequence[str], channel: str, T: int) -> np.ndarray:
-    """Собирает матрицу [n_series, T] по указанному каналу с NaN.
-    """
-    rows = []
-    for w in wells:
-        g = panel_long.loc[panel_long["well_name"] == w, ["t", channel]].sort_values("t")
-        v = np.full(T, np.nan, float)
-        t = g["t"].to_numpy(int)
-        vals = g[channel].to_numpy(float)
-        t = t[(t >= 0) & (t < T)]
-        vals = vals[: len(t)]
-        v[t[: len(vals)]] = vals
-        rows.append(v)
-    return np.vstack(rows) if rows else np.empty((0, T))
-
 
 def _barycenter_or_median(M: np.ndarray, method: str = "auto", gamma: float = 1.0, max_iter: int = 50) -> np.ndarray:
     """Возвращает барицентр (soft-DTW/DBA) или медиану по времени, если tslearn недоступен.
